+++ app/Support/Csp/ContentPolicy.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Support\Csp;
+
+use Spatie\Csp\Policies\Basic;
+
+class ContentPolicy extends Basic
+{
+    public function configure()
+    {
+        parent::configure();
+    }
+}